Move client.py to _internal
