--- docs/api/client.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.client`
--------------------------
-
-.. automodule:: certbot.client
-   :members:
